change user types
